Shifting start UI logic out from MainActivity to PNR_Prompt
